+++ test/tape/seed.js
@@ -6,7 +6,7 @@ var Seed = require('../../lib/seed/index.js');
 tape('checks config.seeds for seed config', function(t) {
   t.plan(1);
   var seeder = new Seed({
-    client: { config: { seeds: { directory: '/some/dir' } } },
+    client: { config: { seeds: { globPatterns: '/some/dir/*.js' } } },
   });
-  t.equal(seeder.config.directory, '/some/dir');
+  t.equal(seeder.config.globPatterns, '/some/dir/*.js');
 });